requirements.txt
aiogram>=3.15,<4
motor>=3.6,<4
pymongo>=4.10,<5
google-generativeai>=0.8,<1
pydantic>=2.10,<3
apscheduler>=3.10,<4
python-dotenv>=1.0,<2
aiofiles>=24.1,<25
